@@ -4,17 +4,17 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  hive: 1.2.0
-  provider: any
+  hive: 2.2.3
+  provider: 6.1.2
 
 dev_dependencies:
-  pedantic: 1.9.0
+  pedantic: 1.11.0
 
 flutter:
   uses-material-design: true
