@@ -9,14 +9,14 @@ class AppState extends ChangeNotifier {
     notifyListeners();
   }
 
-  String _boxName;
+  late String _boxName;
   String get boxName => _boxName;
   set boxName(String boxName) {
     _boxName = boxName;
     notifyListeners();
   }
 
-  Map<dynamic, dynamic> _entries;
+  late Map<dynamic, dynamic> _entries;
   Map<dynamic, dynamic> get entries => _entries;
   set entries(Map<dynamic, dynamic> entries) {
     _entries = entries;
