@@ -6,7 +6,7 @@ class PathView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Consumer<AppState>(
-      builder: (BuildContext context, AppState app, Widget child) {
+      builder: (BuildContext context, AppState app, Widget? child) {
         return Row(
           children: <Widget>[
             _buildPathElement(app.boxName, () {
@@ -22,8 +22,7 @@ class PathView extends StatelessWidget {
     );
   }
 
-  Widget _buildPathElement(String name, VoidCallback goto,
-      {bool divider = true}) {
+  Widget _buildPathElement(String name, VoidCallback goto, {bool divider = true}) {
     return InkWell(
       onTap: goto,
       child: Text(
